Rename package to org.etherpad.client as etherpad.org is a real DNS domain.
