+++ pom.xml
@@ -2,7 +2,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>org.etherpad_lite_client</groupId>
+  <groupId>org.etherpad</groupId>
   <artifactId>etherpad_lite_client</artifactId>
   <version>1.2.1-SNAPSHOT</version>
   <packaging>jar</packaging>
