+++ src/main/java/org/etherpad/client/EPLiteClient.java
@@ -1,4 +1,4 @@
-package org.etherpad_lite_client;
+package org.etherpad.client;
 
 import java.util.Date;
 import java.util.HashMap;
