+++ src/main/java/org/etherpad/client/Request.java
@@ -1,4 +1,4 @@
-package org.etherpad_lite_client;
+package org.etherpad.client;
 
 public interface Request {
     String send() throws Exception;
